requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioice==0.10.2
aiortc==1.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.75.0
anyascii==0.3.3
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==14.4.0
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
Cython==3.2.3
dateparser==1.1.8
decorator==5.2.1
deepgram-sdk==5.3.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.17.0
einops==0.8.1
encodec==0.1.1
eval_type_backport==0.2.2
exceptiongroup==1.3.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.8.0
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hume==0.13.5
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonlines==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
loguru==0.7.3
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.8.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
narwhals==2.14.0
networkx==2.8.8
nltk==3.9.2
noisereduce==3.0.3
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
openai==2.14.0
openai-whisper==20250625
packaging==25.0
pandas==1.5.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pylibsrtp==0.12.0
pynndescent==0.5.13
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.11.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
streamlit==1.50.0
streamlit-webrtc==0.63.11
SudachiDict-core==20251022
SudachiPy==0.6.10
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.9
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
tornado==6.5.4
tqdm==4.67.1
trainer==0.0.36
transformers==4.57.3
TTS==0.22.0
typeguard==4.4.4
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.6.2
wasabi==1.1.3
weasel==0.4.3
websockets==13.1
Werkzeug==3.1.4
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
